Balík oprav

+++ Header Files/visualizable.h
@@ -13,7 +13,7 @@ public:
 
     virtual void addPointCloud(const std::vector<TrackView>& trackViews) = 0;
 
-    virtual void addCamera(const cv::Matx34d camPose) = 0;
+    virtual void addCamera(const cv::Matx34f  camPose) = 0;
 
     virtual void visualize() = 0;
 };
